knowing stack

+++ src/Analyst.java
@@ -2,7 +2,7 @@ public class Analyst extends Employee{
 
 
     public Analyst(String name, double salary, int age) {
-      super(name,salary,age);
+      super(name, (int) salary,age);
 
     }
     public double getAnnualBonus(){
